Refactor OptionSettings
